enabled way of manually capturing errors in sentry
